231012 hye 11:23

+++ src/main/java/com/example/demo/info/controller/InfoController.java
@@ -31,7 +31,7 @@ public class InfoController {
 	@Autowired
 	private PageDTO pdto;
 	
-	private int currentPage;
+	private long currentPage;
 
 	public InfoController() {}
